+++ src/widgets/icon.rs
@@ -5,7 +5,7 @@ use iced::widget::svg;
 use crate::messages::Message;
 
 #[cfg(feature = "icons")]
-pub fn icon(handle: &svg::Handle, size: u16, color: IconColor) -> Container<'static, Message> {
+pub fn icon(handle: &svg::Handle, size: f32, color: IconColor) -> Container<'static, Message> {
     let svg_icon = svg(handle.clone()).width(size).height(size).style(
         move |theme: &iced::theme::Theme, _status| {
             let color = match color {
